refactor: Simplify index.js by extracting configuration into separate modules and update route mounting in index.js and routes/index.js
